Moving fonts & icons to assets
