+++ dapp/src/assets/fonts/Segment-Bold.otf
