+++ dapp/src/assets/icons/ArrowUpRight.tsx
