+++ dapp/src/assets/icons/ConnectBTCAccount.tsx
