+++ dapp/src/assets/icons/Ethereum.tsx
