+++ dapp/src/assets/icons/index.ts
