+++ dapp/src/components/Header/ConnectWallet.tsx
@@ -8,7 +8,7 @@ import {
 } from "#/hooks"
 import { CurrencyBalance } from "#/components/shared/CurrencyBalance"
 import { TextMd } from "#/components/shared/Typography"
-import { Bitcoin, Ethereum } from "#/static/icons"
+import { Bitcoin, Ethereum } from "#/assets/icons"
 import { truncateAddress, asyncWrapper } from "#/utils"
 
 export type ConnectButtonsProps = {
