+++ dapp/src/components/Header/index.tsx
@@ -1,6 +1,6 @@
 import React from "react"
 import { Flex, HStack, Icon } from "@chakra-ui/react"
-import { AcreLogo } from "#/static/icons"
+import { AcreLogo } from "#/assets/icons"
 import ConnectWallet from "./ConnectWallet"
 
 export default function Header() {
