+++ dapp/src/components/Modals/Support/index.tsx
@@ -4,7 +4,7 @@ import {
   useRequestEthereumAccount,
   useWalletContext,
 } from "#/hooks"
-import { ConnectBTCAccount, ConnectETHAccount } from "#/static/icons"
+import { ConnectBTCAccount, ConnectETHAccount } from "#/assets/icons"
 import MissingAccount from "./MissingAccount"
 
 export default function SupportWrapper({
