+++ dapp/src/components/Overview/PositionDetails.tsx
@@ -12,7 +12,7 @@ import {
 } from "@chakra-ui/react"
 import { CurrencyBalanceWithConversion } from "#/components/shared/CurrencyBalanceWithConversion"
 import { TextMd } from "#/components/shared/Typography"
-import { Info } from "#/static/icons"
+import { Info } from "#/assets/icons"
 import StakingModal from "../Modals/Staking"
 
 export default function PositionDetails(props: CardProps) {
