+++ dapp/src/components/Overview/index.tsx
@@ -2,7 +2,7 @@ import React from "react"
 import { Button, Flex, Grid, HStack, Icon, Switch } from "@chakra-ui/react"
 import { useDocsDrawer } from "#/hooks"
 import { TextSm } from "#/components/shared/Typography"
-import { ArrowUpRight } from "#/static/icons"
+import { ArrowUpRight } from "#/assets/icons"
 import { USD } from "#/constants"
 import PositionDetails from "./PositionDetails"
 import Statistics from "./Statistics"
